+++ lib/src/extensions/focus.dart
@@ -5,4 +5,16 @@ extension FocusModel on Focus {
   String getId() {
     return slugify("focus-$name");
   }
+
+  Attribute getAssociatedAttribute() {
+    const attributes = [
+      Attributes.agility,
+      Attributes.charisma,
+      Attributes.dexterity,
+      Attributes.perception,
+      Attributes.strength,
+      Attributes.wisdom
+    ];
+    return attributes.where((element) => element.focuses.contains(this)).first;
+  }
 }