+++ test/src/models/content/markdown_content_test.dart
@@ -15,7 +15,18 @@ This is a paragraph
 - list
 """).toHtml();
       String expected =
-          '<h1>Test</h1><p>This is a paragraph</p><ul><li>This</li><li>is</li><li>a</li><li>list</li></ul>';
+          '<header id="test"><h1>Test</h1></header><p>This is a paragraph</p><ul><li>This</li><li>is</li><li>a</li><li>list</li></ul>';
+
+      expect(actual, equals(expected));
+    });
+    test('markdown headers become headings', () {
+      String actual = MarkdownContent(markdown: """
+# Test 1
+## Test 2
+### Test 3
+""").toHtml();
+      String expected =
+          '<header id="test-1"><h1>Test 1</h1></header><header id="test-2"><h2>Test 2</h2></header><header id="test-3"><h3>Test 3</h3></header>';
 
       expect(actual, equals(expected));
     });